requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.1
requests==2.32.3
python-dotenv==1.0.1
aiofiles==23.2.1
gemini-srt-translator